shipping-service/requirements.txt
django
djangorestframework
requests
django-cors-headers
dj-database-url
gunicorn
